refactor: extract canvas command runtime into tau-ops
