+++ crates/tau-coding-agent/Cargo.toml
@@ -47,14 +47,12 @@ reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
-shell-words.workspace = true
 tokio.workspace = true
 tokio-stream.workspace = true
 tokio-tungstenite.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 wait-timeout.workspace = true
-yrs.workspace = true
 wasmparser.workspace = true
 rustyline.workspace = true
