+++ crates/tau-ops/src/lib.rs
@@ -1,3 +1,4 @@
+mod canvas_commands;
 mod channel_store_admin;
 mod daemon_runtime;
 mod macro_commands;
@@ -5,6 +6,7 @@ mod project_index;
 mod qa_loop_commands;
 mod transport_health;
 
+pub use canvas_commands::*;
 pub use channel_store_admin::*;
 pub use daemon_runtime::*;
 pub use macro_commands::*;
